The final code of the PySparkMl module
The final code of the PySparkMl module

Step 1) Read the data from either HDFS or Local and load the dataframe
Step 2) 
a) For the feature “balance”, whenever the balance was less than 0 we made it 0 else kept it as it is
b) For the feature “pdays”, whenever pdays was less than 0 we made it 0 else keep it as it is
c) The target column name is “y” , changed the column name to “Approved”. Wherever there was a value of “no” we made it 0
and when value was “yes” we made it 1

Step 3) We came up with several interesting visualizations to see the effect of different features for the target variable. For example we see how different kind of job types effects of approval on loan or term deposit. We drew as many visualizations as possible to gain better in-sight of data

Step 4) We checked for class imbalance of the target value in the data

Step 5) We performed the necessary feature extraction & transformations to get the feature vector on which we applied different Spark ML models

Step 6) We later analyzed the results